Remove base ruby env during cleanup to prevent gem version issues (#97)
Currently, our pre-build formula cleanup step removes all preinstalled formula from the macOS host, but it doesn't remove some leftover ruby env files.

These leftover files may cause issues when the cache isn't used (eg. when we change the builder setup, which causes a cache miss) - in these cases, there can be issues if the gems we install don't match the gem versions that are left over in the ruby env.

For the sake of build reproducibility & not having builds break every time we want to update something in the builder just because the base image was updated, this PR completely removes the leftover ruby env, allowing us to install whichever gem versions we want, and allowing us to update gems only when we want to.